# [V1] Simpify vision block hash for prefix caching by removing offset from hash