Reuse the example in https://github.com/vllm-project/vllm/pull/11187

Taking a series of 3 blocks as an example: T0,T1,P00,P01 | P02,P03,P04,T2 | T3,P10,P11,P12, where Ti is i-th text token and Pxy is the y-th placeholder token of the x-th image, so this prompt has 2 images (P0 and P1). Assuming the image hash of P0 and P1 is aaa and bbb, respectively, and mm_positions=[(offset=2, length=5), (offset=9, length=3)], the hash of 3 blocks in https://github.com/vllm-project/vllm/pull/11187 is as follows
```
# (Parent hash,
#  token ID w. placeholders,
#  image hash, start)
hash0 = hash(None, T0,T1,P00,P01, (aaa,0))
hash1 = hash(hash0, P02,P03,P04,T2, (aaa,2))
hash2 = hash(hash1, T3,P10,P11,P12, (bbb,0))
```

But offset is redundant and the following hash format is enough:
```
# (Parent hash,
#  token ID w. placeholders,
#  image hash)
hash0 = hash(None, T0,T1,P00,P01, aaa)
hash1 = hash(hash0, P02,P03,P04,T2, aaa)
hash2 = hash(hash1, T3,P10,P11,P12, bbb)
```

Simple proof:
We need to distinguish the above example with the following cases:
1. The same image but different offset, e.g., T0,T1,T2,P00 | P01,P02,P03,P04 | T3,P10,P11,P12. The hash0 becomes hash(None, T0,T1, **T2,P00**, aaa), which is different
2. The same offset but different image, e.g.,  T0,T1,P20,P21 | P22,P23,P24,T2 | T3,P10,P11,P12. The hash0 becomes hash(None, T0,T1,P20,P21, **ccc**), which is different

This hash format has almost the same speed as previous one, but can avoid the confusion about why we need to add offset to the hash.

Benchmark result on H100 with the script in https://github.com/vllm-project/vllm/pull/11187: https://gist.github.com/comaniac/ea26df17fdffa533cf53d53b8455bc31 I ran 3 times for each commit
```
VLLM_USE_V1=1 VLLM_ENABLE_V1_MULTIPROCESSING=1 python3 mmmu_bench.py --model llava-hf/llava-v1.6-mistral-7b-hf --num-prompts 500  --image-hit-rate 0.3
Without this pr:
Throughput: 15.84 req/s
Throughput: 16.09 req/s
Throughput: 16.00 req/s

With this pr:
Throughput: 16.11 req/s
Throughput: 15.94 req/s
Throughput: 15.84 req/s
```

Note that the following format that only adds the image hash to the first block of this image is also correct, but is a little slower (15.50 reqs/s). The code is here https://github.com/vllm-project/vllm/commit/539e84c58f8ac08d8764573f402057e58ca6d7d2
```
# (Parent hash,
#  token ID w. placeholders,
#  image hash)
hash0 = hash(None, T0,T1,P00,P01, aaa)
hash1 = hash(hash0, P02,P03,P04,T2, )
hash2 = hash(hash1, T3,P10,P11,P12, bbb)
```