@ywang96 2024-12-31
I guess we previously overlooked the fact that the offset information can be already inferred from the token sequence and the hash of the previous block. Thanks for this simplification!